Move all project geometry types to a types folder;
Includes updates to files with relative paths
