+++ app/controllers/projects/show.js
@@ -1,10 +1,10 @@
 import Controller from '@ember/controller';
 import { action } from '@ember-decorators/object';
 import { EmptyFeatureCollection } from '../../models/project';
-import { developmentSiteLayer } from '../../components/project-geometries/development-site';
-import { projectAreaLayer } from '../../components/project-geometries/project-area';
-import { rezoningAreaLayer } from '../../components/project-geometries/rezoning-area';
-import { underlyingZoningLayer, underlyingZoningLabelsLayer } from '../../components/project-geometries/underlying-zoning';
+import { developmentSiteLayer } from '../../components/project-geometries/types/development-site';
+import { projectAreaLayer } from '../../components/project-geometries/types/project-area';
+import { rezoningAreaLayer } from '../../components/project-geometries/types/rezoning-area';
+import { underlyingZoningLayer, underlyingZoningLabelsLayer } from '../../components/project-geometries/types/underlying-zoning';
 import {
   commercialOverlaysLayer,
   coLayer,
@@ -18,8 +18,8 @@ import {
   c23Layer,
   c24Layer,
   c25Layer,
-} from '../../components/project-geometries/commercial-overlays';
-import { specialPurposeDistrictsLayer, specialPurposeDistrictsLabelsLayer } from '../../components/project-geometries/special-purpose-districts';
+} from '../../components/project-geometries/types/commercial-overlays';
+import { specialPurposeDistrictsLayer, specialPurposeDistrictsLabelsLayer } from '../../components/project-geometries/types/special-purpose-districts';
 import areaMapLegendConfig from '../../utils/area-map-legend-config';
 
 export default class ShowProjectController extends Controller {
