+++ app/templates/components/project-geometries.hbs
@@ -33,7 +33,7 @@
       {{/map.labs-layers}}
 
       {{!-- invoke the component for the specified type --}}
-      {{component (concat 'project-geometries/' type)
+      {{component (concat 'project-geometries/types/' type)
         model=model
         mode=mode
         map=map}}
